backend/pyproject.toml
[project]
name = "codemind-backend"
version = "0.1.0"
description = "CodeMind agentic codebase assistant — backend"
requires-python = ">=3.11"

dependencies = [
    # Web framework — handles HTTP requests and routing
    "fastapi>=0.111.0",

    # ASGI server — runs the FastAPI app
    "uvicorn[standard]>=0.29.0",

    # LangChain core — base abstractions for LLMs, prompts, chains
    "langchain>=0.2.0",

    # LangChain + OpenAI integration — lets LangChain talk to GPT models
    "langchain-openai>=0.1.0",

    # LangChain + Weaviate integration — lets LangChain talk to our vector DB
    "langchain-weaviate>=0.0.3",

    # LangGraph — builds stateful multi-agent workflows as graphs
    "langgraph>=0.1.0",

    # LangSmith — observability, tracing, and evals for LangChain apps
    "langsmith>=0.1.0",

    # Weaviate Python client — connects to our local Weaviate instance
    "weaviate-client>=4.6.0",

    "tree-sitter>=0.23.0",

    # Official per-language grammars (Python 3.13 compatible)
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",

    # Reads .env file and loads it into environment variables
    "python-dotenv>=1.0.0",

    # Data validation — used to define typed config and request/response schemas
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",

    # Async HTTP client — for streaming responses to the frontend
    "httpx>=0.27.0",

    # Handles multipart file uploads (e.g. uploading a zip of a codebase)
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]